+++ backend/api/__test__/sampletest.test.ts
@@ -1,10 +1,10 @@
 import request from 'supertest';
-import app from '../app';
+import index from '../index';
 
 describe('API Endpoints', () => {
   describe('GET /', () => {
     it('should return 200 OK with hello message', async () => {
-      const response = await request(app).get('/');
+      const response = await request(index).get('/');
       expect(response.status).toBe(200);
       expect(response.body).toHaveProperty('message');
       expect(response.body.message).toBe('API is running');
@@ -13,7 +13,7 @@ describe('API Endpoints', () => {
 
   describe('Error handling', () => {
     it('should return 404 for undefined routes', async () => {
-      const response = await request(app).get('/nonexistent-route');
+      const response = await request(index).get('/nonexistent-route');
       expect(response.status).toBe(404);
     });
   });
